% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_box_function_import_count_linter.R
\name{box_func_import_count_linter}
\alias{box_func_import_count_linter}
\title{Box library function import count linter}
\usage{
box_func_import_count_linter(max = 8L)
}
\arguments{
\item{max}{Maximum function imports allowed between \code{[} and \verb{]}. Defaults to 8.}
}
\description{
Checks that function imports do not exceed the defined \code{max}.
}
\examples{
# will produce lints
lintr::lint(
  text = "box::use(package[one, two, three, four, five, six, ])",
  linters = box_func_import_count_linter()
)

lintr::lint(
  text = "box::use(package[one, two, three, four, ])",
  linters = box_func_import_count_linter(3)
)

# okay
lintr::lint(
  text = "box::use(package[one, two, three, four, five, ])",
  linters = box_func_import_count_linter()
)

lintr::lint(
  text = "box::use(package[one, two, three, ])",
  linters = box_func_import_count_linter(3)
)

}
