% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_box_trailing_commas_linter.R
\name{box_trailing_commas_linter}
\alias{box_trailing_commas_linter}
\title{Box library trailing commas linter}
\usage{
box_trailing_commas_linter()
}
\description{
Checks that all \code{box:use} imports have a trailing comma. This applies to
package or module imports between \code{(} and \verb{)}, and function imports between
\code{[} and \verb{]}. Take note that \code{lintr::commas_linter()} may come into play.
}
