% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_box_universal_import_linter.R
\name{box_universal_import_linter}
\alias{box_universal_import_linter}
\title{Box library universal import linter}
\usage{
box_universal_import_linter(exclude = c("shiny", "testthat"))
}
\arguments{
\item{exclude}{Character vector of package names to exclude. Defaults are \code{shiny} and \code{testthat}.}
}
\description{
Checks that all function imports are explicit. \code{package[...]} is not used.
}
\examples{
# will produce lints
lintr::lint(
  text = "box::use(base[...])",
  linters = box_universal_import_linter()
)

lintr::lint(
  text = "box::use(path/to/file[...])",
  linters = box_universal_import_linter()
)

# okay
lintr::lint(
  text = "box::use(base[print])",
  linters = box_universal_import_linter()
)

lintr::lint(
  text = "box::use(path/to/file[do_something])",
  linters = box_universal_import_linter()
)

}
