% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageserver.R
\name{box_use_parser}
\alias{box_use_parser}
\title{Add \verb{\{languageserver\}`` parsers for }{box}`}
\usage{
box_use_parser(expr, action)
}
\value{
None. This function is called for its side effects.
}
\description{
Add \verb{\{languageserver\}`` parsers for }{box}`
}
