% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{build_sass}
\alias{build_sass}
\title{Build Sass}
\usage{
build_sass(watch = TRUE)
}
\arguments{
\item{watch}{Keep the process running and rebuilding Sass whenever source files change.
Only supported for \code{sass: node} config in \code{rhino.yml}.}
}
\description{
Build Sass
}
