% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react.R
\name{$,RhinoReact-method}
\alias{$,RhinoReact-method}
\title{Rhino React}
\usage{
\S4method{$}{RhinoReact}(x, name)
}
\arguments{
\item{x}{The Rhino React singleton.}

\item{name}{The name of the React component.}
}
\description{
Rhino React
}
