% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{covr_report}
\alias{covr_report}
\title{Display rhino test coverage results using a standalone report}
\usage{
covr_report(rhino_coverage = covr_r(), ...)
}
\arguments{
\item{rhino_coverage}{a rhino coverage dataset, defaults to \code{covr_r()}.}

\item{...}{additional arguments to pass to
\href{https://covr.r-lib.org/reference/report.html}{\code{covr::report()}}}
}
\value{
None. This function is called for side effects.
}
\description{
Uses the \code{{covr}} package to produce unit test coverage reports.
Uses the \code{{testhat}} package to run all unit tests in \code{tests/testthat} directory.
}
\examples{
if (interactive()) {
  # Run a test coverage report on a rhino app
  covr_report()
}

}
