% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependency_add}
\alias{dependency_add}
\title{Add (install) packages to Rhino app project}
\usage{
dependency_add(packages)
}
\arguments{
\item{packages}{character vector of package names}
}
\value{
None. This function is called for side effects.
}
\description{
Adds the requested packages to the 'dependencies.R' file, installs them and does a {renv}
snapshot in an already initialized Rhino app.
}
