% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependency_remove}
\alias{dependency_remove}
\title{Remove (uninstall) packages from Rhino app project}
\usage{
dependency_remove(packages)
}
\arguments{
\item{packages}{character vector of package names}
}
\value{
None. This function is called for side effects.
}
\description{
Remove the requested packages from the 'dependencies.R' file, uninstalls them and does a {renv}
snapshot in an already initialized Rhino app.
}
