% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhino.R
\name{diagnostics}
\alias{diagnostics}
\title{Print diagnostics}
\usage{
diagnostics()
}
\value{
None. This function is called for side effects.
}
\description{
Prints information which can be useful for diagnosing issues with Rhino.
}
\examples{
if (interactive()) {
  # Print diagnostic information.
  diagnostics()
}
}
