% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{format_r}
\alias{format_r}
\title{Format R}
\usage{
format_r(paths)
}
\arguments{
\item{paths}{Character vector of files and directories to format.}
}
\value{
None. This function is called for side effects.
}
\description{
Uses the \code{{styler}} package to automatically format R sources.
}
