% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{format_sass}
\alias{format_sass}
\title{Format Sass}
\usage{
format_sass(fix = TRUE)
}
\arguments{
\item{fix}{If \code{TRUE}, fixes formatting. If FALSE, reports formatting errors without fixing them.}
}
\value{
None. This function is called for side effects.
}
\description{
Runs \href{https://prettier.io/}{prettier} on Sass (.scss) files in \code{app/styles} directory.
Requires Node.js installed.
}
\details{
You can prevent prettier from formatting a given chunk of your code by adding a special comment:

\if{html}{\out{<div class="sourceCode scss">}}\preformatted{// prettier-ignore
}\if{html}{\out{</div>}}

Read more about \href{https://prettier.io/docs/en/ignore}{ignoring code}.
}
