% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHINO.R
\name{getKeywords}
\alias{getKeywords}
\title{getKeywords}
\usage{
getKeywords(data, n = 10, length = 1, pos = "noun")
}
\arguments{
\item{data}{vector. Source data for keywords}

\item{n}{integer. Maximum number of keywords. Default is 10}

\item{length}{Minimum length of character of every keyword. Default is 1}

\item{pos}{The Part-Of-Speech type you want to extract. Default is "noun"}
}
\value{
Vector of extracted keywords
}
\description{
Get keywords.
}
\examples{
getKeywords("Input Korean sentences here here here", n=3, length = 2, pos ="ALL")
}
