% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHINO.R
\name{getMorph}
\alias{getMorph}
\title{getMorph}
\usage{
getMorph(sentence, type = "noun", file = FALSE)
}
\arguments{
\item{sentence}{Korean sentences. Try Korean words or sentences.}

\item{type}{The Part-Of-Speech type you want to extract. noun(NNG, NNP, NP), verb(VV, VA, XR), NNG, NNP, NP, NNB, VV, VA, XR, VX. Default is noun.}

\item{file}{If file itself is input (currently not realized. Use lapply()). See the example at my blog(http://blog.naver.com/lingua/220848428864)}
}
\value{
vector of extracted morph result
}
\description{
Get analyzed Korean morph. The initRhino() must be executed once the first.
}
\examples{
initRhino()
getMorph("Input Korean sentences here.", "NNP")
}

