% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{lint_r}
\alias{lint_r}
\title{Lint R}
\usage{
lint_r(accepted_errors = 0)
}
\arguments{
\item{accepted_errors}{Number of accepted style errors.}
}
\value{
None. This function is called for side effects.
}
\description{
Uses the \code{{lintr}} package to check all R sources in the \code{app} and \code{tests/testthat} directories
for style errors.
}
