% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{lint_sass}
\alias{lint_sass}
\title{Lint Sass}
\usage{
lint_sass(fix = FALSE)
}
\arguments{
\item{fix}{Automatically fix problems.}
}
\value{
None. This function is called for side effects.
}
\description{
Runs \href{https://stylelint.io/}{Stylelint} on the Sass sources in the \code{app/styles} directory.
Requires Node.js and the \code{yarn} command to be available on the system.
}
\examples{
if (interactive()) {
  # Lint the Sass sources in the `app/styles` directory.
  lint_sass()
}
}
