% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react.R
\docType{data}
\name{react}
\alias{react}
\title{React components}
\format{
An object of class \code{RhinoReact} of length 1.
}
\usage{
react
}
\description{
Access the React components defined in your app.
}
\examples{
react$MyComponent("Hello!")

}
\keyword{datasets}
