% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rhinos}
\alias{rhinos}
\title{Population of rhinos}
\format{
A data frame with 58 rows and 3 variables:
\describe{
\item{Year}{year}
\item{Population}{rhinos population}
\item{Species}{rhinos species}
}
}
\source{
\url{https://ourworldindata.org/}
}
\usage{
rhinos
}
\description{
A dataset containing population of 5 species of rhinos.
}
\keyword{datasets}
