% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{say_hello}
\alias{say_hello}
\title{Say hello}
\usage{
say_hello()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Says hello.
}
\details{
Use this function to say hello.
}
