% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{test_e2e}
\alias{test_e2e}
\title{Run Cypress end-to-end tests}
\usage{
test_e2e(interactive = FALSE)
}
\arguments{
\item{interactive}{Should Cypress be run in the interactive mode?}
}
\description{
Run Cypress end-to-end tests
}
