% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{use_e2e_tests}
\alias{use_e2e_tests}
\title{Add end-to-end tests}
\usage{
use_e2e_tests()
}
\value{
None. This function is called for side effects.
}
\description{
Adds the Rhino Cypress end-to-end test structure to a Rhino application:
the \code{tests/cypress.config.js} file and the \verb{tests/cypress/} directory.
}
\details{
This structure is added automatically by \code{\link[=init]{init()}}.
Use this function to add it to an existing Rhino project.

If \code{tests/cypress.config.js} or the \verb{tests/cypress/} directory already exist
in an interactive session, you will be prompted to either abort (the default)
or back up the existing paths (each moved to a \code{.bak} path) and create new
ones. In non-interactive sessions the function aborts with an error.
}
\examples{
if (interactive()) {
  # Add the end-to-end test structure to the current Rhino project.
  use_e2e_tests()
}
}
