% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{covr_r}
\alias{covr_r}
\title{Run a unit test coverage check}
\usage{
covr_r(
  source_files = list.files("app", pattern = "\\\\.[rR]$", full.names = TRUE, recursive =
    TRUE),
  test_files = list.files("tests/testthat", pattern = "^test-.*\\\\.R", full.names =
    TRUE, recursive = TRUE),
  line_exclusions = NULL,
  function_exclusions = NULL
)
}
\arguments{
\item{source_files}{Character vector of source files with function definitions to measure coverage.
Defaults to all \code{.R} files in the \code{app} tree.}

\item{test_files}{Character vector of test files with code to test the functions. Defaults to
all test files in \code{tests/testthat} with the \verb{test-<name>.R} filename pattern.}

\item{line_exclusions}{passed to \code{covr::file_coverage}}

\item{function_exclusions}{passed to \code{covr::file_coverage}}
}
\value{
A \code{covr} coverage dataset.
}
\description{
Uses the \code{{covr}} package to produce unit test coverage reports.
Uses the \code{{testhat}} package to run all unit tests in \code{tests/testthat} directory.
}
\examples{
if (interactive()) {
  # Run a test coverage check for the entire rhino app
  # using all tests in the `tests/testthat` directory.
  covr_r()
}

}
