% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{covr_r}
\alias{covr_r}
\title{Run a test coverage check}
\usage{
covr_r()
}
\value{
A \code{covr} coverage dataset.
}
\description{
Uses the \code{{covr}} package to produce unit test coverage reports.
Uses the \code{{testhat}} package to run all unit tests in \code{tests/testthat} directory.
}
\examples{
if (interactive()) {
  # Run a test coverage check for the entire rhino app 
  # using all tests in the `tests/testthat` directory.
  covr_r()
}

}
