% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{format_r}
\alias{format_r}
\title{Format R}
\usage{
format_r(paths)
}
\arguments{
\item{paths}{Character vector of files and directories to format.}
}
\value{
None. This function is called for side effects.
}
\description{
Uses the \code{{styler}} package to automatically format R sources.
}
\details{
The code is formatted according to the \code{styler::tidyverse_style} guide with one adjustment:
spacing around math operators is not modified to avoid conflicts with \code{box::use()} statements.
}
\examples{
if (interactive()) {
  # Format a single file.
  format_r("app/main.R")

  # Format all files in a directory.
  format_r("app/view")
}

}
