% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Create Shiny application using \code{{rhino}}}
\usage{
init(dir = ".", github_actions_ci = TRUE)
}
\arguments{
\item{dir}{Name of the directory to create application in.}

\item{github_actions_ci}{Should the Github Actions CI be added.}
}
\description{
Create Shiny application using \code{{rhino}}
}
