% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Create Shiny application using \code{{rhino}}}
\usage{
init(dir = ".", github_actions_ci = TRUE, rhino_install_source = "rhino")
}
\arguments{
\item{dir}{Name of the directory to create application in.}

\item{github_actions_ci}{Should the Github Actions CI be added.}

\item{rhino_install_source}{Passed to \code{renv::install()} when using an existing \code{renv.lock}.}
}
\description{
Create Shiny application using \code{{rhino}}
}
