% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{lint_r}
\alias{lint_r}
\title{Lint R}
\usage{
lint_r(paths = NULL)
}
\arguments{
\item{paths}{Character vector of directories and files to lint.
When \code{NULL} (the default), check \code{app} and \code{tests/testthat} directories.}
}
\value{
None. This function is called for side effects.
}
\description{
Uses the \code{{lintr}} package to check all R sources in the \code{app} and \code{tests/testthat} directories
for style errors.
}
\details{
The linter rules can be \href{https://lintr.r-lib.org/articles/lintr.html#configuring-linters}{adjusted}
in the \code{.lintr} file.

You can set the maximum number of accepted style errors
with the \code{legacy_max_lint_r_errors} option in \code{rhino.yml}.
This can be useful when inheriting legacy code with multiple styling issues.

The \code{\link[box.linters:namespaced_function_calls]{box.linters::namespaced_function_calls()}} linter requires the \code{{treesitter}} and
\code{{treesitter.r}} packages. These require R >= 4.3.0. \code{lint_r()} will continue to run and skip
\code{namespaced_function_calls()} if its dependencies are not available.
}
