% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{lint_r}
\alias{lint_r}
\title{Lint R}
\usage{
lint_r(accepted_errors = 0)
}
\arguments{
\item{accepted_errors}{Number of accepted style errors.}
}
\value{
None. This function is called for side effects.
}
\description{
Uses the \code{{lintr}} package to check all R sources in the \code{app} and \code{tests/testthat} directories
for style errors.
}
\details{
The linter rules can be adjusted in the \code{.lintr} file.
}
\examples{
if (interactive()) {
  # Lint all R sources in the `app` and `tests/testthat` directories.
  lint_r()

  # Accept up to 10 errors:
  lint_r(accepted_errors = 10)
}

}
