% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{lint_r}
\alias{lint_r}
\title{Lint R}
\usage{
lint_r()
}
\value{
None. This function is called for side effects.
}
\description{
Uses the \code{{lintr}} package to check all R sources in the \code{app} and \code{tests/testthat} directories
for style errors.
}
\details{
The linter rules can be adjusted in the \code{.lintr} file.

You can set the maximum number of accepted style errors
with the \code{legacy_max_lint_r_errors} option in \code{rhino.yml}.
This can be useful when inheriting legacy code with multiple styling issues.
}
