% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{test_e2e}
\alias{test_e2e}
\title{Run Cypress end-to-end tests}
\usage{
test_e2e(interactive = FALSE)
}
\arguments{
\item{interactive}{Should Cypress be run in the interactive mode?}
}
\value{
None. This function is called for side effects.
}
\description{
Uses \href{https://www.cypress.io/}{Cypress} to run end-to-end tests
defined in the \code{tests/cypress} directory.
Requires Node.js to be available on the system.
}
\details{
If you want to write end-to-end tests with \code{{shinytest2}}, see our
\href{https://appsilon.github.io/rhino/articles/how-to/use-shinytest2.html}{How-to: Use shinytest2}
guide.
}
\examples{
if (interactive()) {
  # Run the end-to-end tests in the `tests/cypress` directory.
  test_e2e()
}
}
