% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{test_r}
\alias{test_r}
\title{Run R unit tests}
\usage{
test_r(
  paths = fs::dir_ls("tests/testthat/", glob = "*.R", recurse = TRUE, type = "file"),
  inline_issues = FALSE,
  raw_testthat_output = FALSE
)
}
\arguments{
\item{paths}{A character vector of paths to R files or directories containing tests.
Given a directory, R files in the directory will be included as test files.
Defaults to all files in all directories recursively in \code{tests/testthat}.}

\item{inline_issues}{If \code{TRUE}, test failure, wanring, and skip messages are shown while the
tests are running. If \code{FALSE}, test failure, warning, and skip messages are shown after
all tests are run.}

\item{raw_testthat_output}{boolean, See return value.}
}
\value{
If \code{raw_testthat_output = FALSE}, a data.frame (invisibly) containing data
about the \code{testthat} test results.
If \code{raw_testthat_output = TRUE}, a list (invisibly) of lists containing data
returned by \code{testthat::test_file()}.
}
\description{
Uses the \code{{testhat}} package to run all unit tests in the \code{tests/testthat} directory.
Alternatively, a vector of paths (files and directories) can be provided.
}
\examples{
if (interactive()) {
  # Run all unit tests in the `tests/testthat` directory, recursively.
  test_r()

  # Run all unit tests in the `tests/testthat` directory only. Non-recursive.
  test_r("tests/testthat")

  # Run one unit test.
  test_r("tests/testthat/main.R")

  # Run unit tests on a collection of files and directories.
  test_r(c("tests/testthat/test-main.R", "tests/testthat/logic"))
}
}
