% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{test_r}
\alias{test_r}
\title{Run R unit tests}
\usage{
test_r()
}
\value{
None. This function is called for side effects.
}
\description{
Uses the \code{{testhat}} package to run all unit tests in \code{tests/testthat} directory.
}
\examples{
if (interactive()) {
  # Run all unit tests in the `tests/testthat` directory.
  test_r()
}

}
